refactor(data): move Config to theme package
it is a theme config file parser indeed.
