+++ app/src/main/java/com/osfans/trime/ime/core/Trime.java
@@ -61,8 +61,8 @@ import com.osfans.trime.BuildConfig;
 import com.osfans.trime.R;
 import com.osfans.trime.core.Rime;
 import com.osfans.trime.data.AppPrefs;
-import com.osfans.trime.data.Config;
 import com.osfans.trime.data.db.draft.DraftHelper;
+import com.osfans.trime.data.theme.Config;
 import com.osfans.trime.databinding.CompositionRootBinding;
 import com.osfans.trime.databinding.InputRootBinding;
 import com.osfans.trime.ime.broadcast.IntentReceiver;
