+++ app/src/main/java/com/osfans/trime/ime/symbol/CandidateAdapter.java
@@ -12,7 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.recyclerview.widget.RecyclerView;
 import com.osfans.trime.core.Rime;
-import com.osfans.trime.data.Config;
+import com.osfans.trime.data.theme.Config;
 import com.osfans.trime.databinding.LiquidKeyItemBinding;
 import java.util.ArrayList;
 import java.util.List;
