+++ app/src/main/java/com/osfans/trime/ime/text/TextInputManager.kt
@@ -10,7 +10,7 @@ import androidx.appcompat.app.AlertDialog
 import com.osfans.trime.R
 import com.osfans.trime.core.Rime
 import com.osfans.trime.data.AppPrefs
-import com.osfans.trime.data.Config
+import com.osfans.trime.data.theme.Config
 import com.osfans.trime.databinding.InputRootBinding
 import com.osfans.trime.ime.broadcast.IntentReceiver
 import com.osfans.trime.ime.core.EditorInstance
